support for glpi not on server root (e.g. http://zzz.fr/glpi/)
