+++ tests/functional/Glpi/Security/ReAuth/ReAuthManagerTest.php
@@ -37,7 +37,6 @@ namespace tests\units\Glpi\Security\ReAuth;
 use Computer;
 use Glpi\Exception\RedirectException;
 use Glpi\Security\ReAuth\ReAuthManager;
-use Glpi\Security\ReAuth\ReAuthStrategyInterface;
 use Glpi\Tests\DbTestCase;
 use Glpi\Tests\Glpi\Security\ReAuth\ReAuthTrait;
 use InvalidArgumentException;
@@ -282,45 +281,5 @@ class ReAuthManagerTest extends DbTestCase
 
         // --- act + assert : resolution falls back to the native Password strategy ---
         $this->assertSame('Password', $manager->getLabel());
-    }
-
-    /**
-     * Build a throwaway strategy with a controllable label, priority and availability.
-     * @todo a déplacer dans le trait dédié
-     */
-    private function makeStrategy(string $label, int $priority, bool $available): ReAuthStrategyInterface
-    {
-        return new readonly class ($label, $priority, $available) implements ReAuthStrategyInterface {
-            public function __construct(
-                private string $label,
-                private int $priority,
-                private bool $available,
-            ) {}
-
-            public function verify(int $users_id, string $user_input): bool
-            {
-                return true;
-            }
-
-            public function isAvailable(int $users_id, int $entities_id = 0): bool
-            {
-                return $this->available;
-            }
-
-            public function getLabel(): string
-            {
-                return $this->label;
-            }
-
-            public function getPromptTemplate(): string
-            {
-                return 'pages/reauth/password_form.html.twig';
-            }
-
-            public function getPriority(): int
-            {
-                return $this->priority;
-            }
-        };
     }
 }
