+++ tests/src/Glpi/Security/ReAuth/ReAuthTrait.php
@@ -35,6 +35,7 @@
 namespace Glpi\Tests\Glpi\Security\ReAuth;
 
 use Glpi\Security\ReAuth\ReAuthManager;
+use Glpi\Security\ReAuth\ReAuthStrategyInterface;
 
 /**
  * Helpers shared by the re-authentication tests: fake a web request context and
@@ -59,6 +60,9 @@ trait ReAuthTrait
     /**
      * Simulate a web (non-CLI) request context so that re-authentication
      * redirects can be triggered from a test.
+     *
+     * The superglobals populated here are the ones ReAuthManager::redirectToReauth()
+     * reads to record the request that triggered the re-authentication.
      *
      * @param array<string, string> $get
      * @param array<string, string> $post
@@ -107,5 +111,44 @@ trait ReAuthTrait
         } else {
             unset($_SESSION['glpi_reauth_until']);
         }
+    }
+
+    /**
+     * Build a throwaway strategy with a controllable label, priority and availability.
+     */
+    private function makeStrategy(string $label, int $priority, bool $available): ReAuthStrategyInterface
+    {
+        return new readonly class ($label, $priority, $available) implements ReAuthStrategyInterface {
+            public function __construct(
+                private string $label,
+                private int $priority,
+                private bool $available,
+            ) {}
+
+            public function verify(int $users_id, string $user_input): bool
+            {
+                return true;
+            }
+
+            public function isAvailable(int $users_id, int $entities_id = 0): bool
+            {
+                return $this->available;
+            }
+
+            public function getLabel(): string
+            {
+                return $this->label;
+            }
+
+            public function getPromptTemplate(): string
+            {
+                return 'pages/reauth/password_form.html.twig';
+            }
+
+            public function getPriority(): int
+            {
+                return $this->priority;
+            }
+        };
     }
 }
